Revert "Bug 2045468 - Make sure to check the correct browsing context. r=smaug" for causing process crashes
This reverts commit 90ab644c07fefbb33aeca75ed2b27f2bca44b7cb.

+++ docshell/base/BrowsingContext.cpp
@@ -4234,7 +4234,7 @@ bool BrowsingContext::CanSet(
   return XRE_IsParentProcess() && !aSource && IsTop();
 }
 
-bool BrowsingContext::CanSet(FieldIndex<IDX_BrowserId>, const uint64_t& aValue,
+bool BrowsingContext::CanSet(FieldIndex<IDX_BrowserId>, const uint32_t& aValue,
                              ContentParent* aSource) {
   // We should only be able to set this for toplevel contexts which don't have
   // an ID yet.
